+++ hybrid_os/hybrid_os_platform_interface/lib/hybrid_os_platform_interface.dart
@@ -1,4 +1,2 @@
-/// hybrid_os_platform_interface.
-library hybrid_os_platform_interface;
-
+export 'src/os_plugin.dart';
 export 'src/os.dart';