+++ hybrid_os/hybrid_os_platform_interface/lib/src/os.dart
@@ -1,31 +1,5 @@
-import 'package:plugin_platform_interface/plugin_platform_interface.dart';
-
-abstract base class OSPlatform extends PlatformInterface implements OS {
-  /// Constructs a [OSPlatform].
-  OSPlatform() : super(token: _token);
-
-  static final Object _token = Object();
-
-  static OSPlatform? _instance;
-
-  /// The default instance of [OSPlatform] to use.
-  static OSPlatform get instance {
-    final instance = _instance;
-    if (instance == null) {
-      throw UnimplementedError('OS is not implemented on this platform.');
-    }
-    return instance;
-  }
-
-  /// Platform-specific implementations should set this with their own
-  /// platform-specific class that extends [OSPlatform] when
-  /// they register themselves.
-  static set instance(OSPlatform instance) {
-    PlatformInterface.verifyToken(instance, _token);
-    _instance = instance;
-  }
-}
+import 'os_plugin.dart';
 
 abstract interface class OS {
-  factory OS() => OSPlatform.instance;
+  factory OS() => OSPlugin.instance.createOS();
 }