+++ hybrid_os/hybrid_os_platform_interface/CHANGELOG.md
@@ -2,3 +2,8 @@
 
 * Add `OS` interface class.
 * Add `OSPlatform` base class.
+
+## 0.1.0-dev.0
+
+* Move `OSPlatform` to `OSPlugin`.
+* Add `OSPlguin#createOS` method.